RegisterIntakeDialog almost finished
